+++ src/valhalla_benchmark_loki.cc
@@ -2,6 +2,7 @@
 
 #include "loki/search.h"
 #include "midgard/logging.h"
+#include "midgard/pointll.h"
 
 #include <algorithm>
 #include <atomic>
@@ -210,7 +211,20 @@ int main(int argc, char** argv) {
     std::ifstream stream(file);
     std::string line;
     while (std::getline(stream, line)) {
-      auto loc = valhalla::baldr::Location::FromCsv(line);
+      // Parse line to get lat,lon
+      std::stringstream ss(line);
+      std::string item;
+      std::vector<std::string> parts;
+      while (std::getline(ss, item, ',')) {
+        parts.push_back(std::move(item));
+      }
+      float lat = std::stof(parts[0]);
+      if (lat < -90.0f || lat > 90.0f) {
+        throw std::runtime_error("Latitude must be in the range [-90, 90] degrees");
+      }
+      float lon = valhalla::midgard::circular_range_clamp<float>(std::stof(parts[1]), -180, 180);
+      valhalla::midgard::PointLL ll(lat, lon);
+      valhalla::baldr::Location loc(ll);
       loc.minimum_reachability_ = isolated;
       loc.radius_ = radius;
       job.emplace_back(std::move(loc));